requirements.txt
asyncpg==0.30.0
fastapi==0.115.12
langchain==0.3.21
langchain_core==0.3.48
langchain_openai==0.3.10
langgraph==0.3.20
psycopg==3.2.4
python-dotenv==1.1.0
Requests==2.32.3
typing_extensions==4.13.0
uvicorn==0.34.0
